Fix random pauses issue#48
This changes the behavior of the visualizer so that when the user clicks
somehwere far away on the progress bar, it will no longer pause the
visualizer.  However the user can still manually pause playback by using
the pause button.
This fixes the random pauses because the visualizer interprets big
changes in the position of the slider as a mouse click, when sometimes
it will move far enough on its own without user interaction.